requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
idna==2.7
IMDbPY==6.6
lxml==4.2.5
pycparser==2.19
PyMySQL==0.9.2
six==1.11.0
SQLAlchemy==1.2.12
